Begin implementing autocomplete
This appends autocomplete text to the search box when a note's title
begins with the typed text, and at the same time selects the note in the
note list, and opens an autocomplete-suggested note when <Enter> is
pressed.

Many of the details and corner-cases of the autocomplete behaviour still
need to be handled.