Some unicode fixes
This is just the start of fixing the unicode handling..

When reading filenames from os.walk() or file contents from files, try
to decode them into unicode, try utf-8 first and if it that doesn't work
various other encodings, including any that chardet comes up with.

This is not perfect, some files may be decoded incorrectly, or decoding
may fail completely if none of the encodings work.

But this fixes crashes when e.g. the user types a non-ascii character
into the search box and it then gets compared against note titles and
contents.

Also encode the unicode filenames to utf-8 in some log messages, and
when passing filenames as arguments to the editor.