Performing more robust shell word splitting.
This should address the uncaught exception when specifying a -e editor option which relies on shell expansion/splitting more complicated than splitting on spaces. It's not clear to me why any encoding check needs to be done here (presumably if the user were to specify a flag at runtime with an odd encoding, they would be using a shell capable of handling said encoding anyway?), but I've left it. Also, the prior code was not re-enconding the base command.